Data Analysis
Here we extracted valuable insights from our data regarding Youth Unemployment in South Africa using the power of Power Bi to show us valuable information that we could use to tell the story of our data and the importance of the model we were building.  This was a very insightful stage.